# docs/openapi.yaml
...
openapi: "3.0.3"
info:
  title: "3API Relay Panel"
  version: "0.5.0"
  description: "Anthropic-compatible LLM relay with multi-tenant storefront, reseller admin, plans / orders / subscriptions, webhook delivery, and per-tenant BYOK upstream channels."
servers:
  - url: "https://3api.pro"
    description: "SaaS root (multi-tenant)"
  - url: "https://{store}.3api.pro"
    description: "Tenant storefront (replace {store})"
  - url: "http://localhost:8080"
    description: "Local dev"
...
components:
  securitySchemes:
    apiKey:
      type: "http"
      scheme: "bearer"
      bearerFormat: "sk-..."
    adminBearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
    customerBearer:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
paths:
...
  /admin/webhooks:
    get:
      summary: "List webhook subscriptions"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "200":
          description: "Webhook list incl. last_triggered_at, fail_count_total"
      security:
        - adminBearer: []
    post:
      summary: "Create webhook subscription"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "201":
          description: "Created (secret returned once)"
      description: "Returns the HMAC secret in the response **once** — record it for signature verification. Subsequent reads will not include the secret. Allowed event types: order.paid, subscription.expired, refund.processed, wholesale.low."
      security:
        - adminBearer: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              url: "https://example.com/3api-webhook"
              events:
                - "order.paid"
                - "subscription.expired"
  /admin/webhooks/{id}:
    patch:
      summary: "Update webhook url / events / enabled"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "200":
          description: "Updated"
      security:
        - adminBearer: []
    delete:
      summary: "Delete webhook (cascades deliveries)"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "200":
          description: "Deleted"
      security:
        - adminBearer: []
  /admin/webhooks/{id}/test:
    post:
      summary: "Send a synthetic test event"
      tags:
        - "admin"
        - "webhooks"
      responses:
        "200":
          description: "Delivery row with status"
      security:
        - adminBearer: []
...